Improve routers
